FEATURE: Add label to Docker images with branch name
To better differentiate between Docker images (given that these have the same name which reflect type of builder) new label is added with name "io.machinekit.machinekit-hal.vcs-branch" which value contains the name of git branch from which the image was built.

This is useful mainly in CI/CD Docker image caching on Github Actions workflow.