FEATURE: Allow building multiarch Docker images for distributions with split repositories
Ubuntu (and probably other distributions) implement something called 'split repos', which means that for different architectures there is different address for basically the same repository.

This commit abstract this away so that the same build rules can be applied to any combination of BUILD and HOST architectures. The repositories become also locked to BUILD and HOST architectures only. This change should help make the builder images more one-functionality tool.